# docs/swagger.yaml
basePath: /
definitions:
  model.BalanceSumaryResponse:
    properties:
      categoryBalance:
        additionalProperties:
          type: number
        type: object
      monthBalance:
        type: number
      totalExpense:
        type: number
      totalReceipt:
        type: number
    type: object
  model.ExpenseRequest:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      value:
        type: number
    type: object
  model.ExpenseResponse:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      userId:
        type: string
      value:
        type: number
    type: object
  model.ReceiptRequest:
    properties:
      date:
        type: string
      description:
        type: string
      value:
        type: number
    type: object
  model.ReceiptResponse:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      userId:
        type: string
      value:
        type: number
    type: object
  model.UserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  title: Alura Backend Challenge 2nd Edition API
  version: 1.0.0
paths:
  /budget-control/api/v1/authenticate:
    post:
      description: authenticate user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      responses:
        "201":
          description: token
          schema:
            type: string
      tags:
      - User
  /budget-control/api/v1/balance/{year}/{month}:
    get:
      description: get month balance sumary
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: string
      - description: Month
        in: path
        name: month
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BalanceSumaryResponse'
      tags:
      - Balance
  /budget-control/api/v1/expense:
    get:
      description: Find all expenses
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ExpenseResponse'
            type: array
      summary: Find all expenses
      tags:
      - Expenses
    post:
      description: Create a new expense
      parameters:
      - description: Expense
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/model.ExpenseRequest'
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a new expense
      tags:
      - Expenses
  /budget-control/api/v1/expense/{id}:
    delete:
      description: Delete an expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
      summary: Delete an expense
      tags:
      - Expenses
    get:
      description: Find expense by id
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ExpenseResponse'
      summary: Find expense by id
      tags:
      - Expenses
    put:
      description: Update an expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      - description: Expense
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/model.ExpenseRequest'
      responses:
        "204":
          description: ""
      summary: Update an expense
      tags:
      - Expenses
  /budget-control/api/v1/expense/{year}/{month}:
    get:
      description: Find all expenses by period
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: integer
      - description: Month
        in: path
        name: month
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ExpenseResponse'
            type: array
      summary: Find all expenses by period
      tags:
      - Expenses
  /budget-control/api/v1/health:
    get:
      description: return server status
      responses:
        "200":
          description: ""
        "404":
          description: ""
      tags:
      - Health
  /budget-control/api/v1/receipt:
    get:
      description: Find all receipts
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ReceiptResponse'
            type: array
      summary: Find all receipts
      tags:
      - Receipts
    post:
      description: Create a new receipt
      parameters:
      - description: Receipt
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/model.ReceiptRequest'
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a new receipt
      tags:
      - Receipts
  /budget-control/api/v1/receipt/{id}:
    delete:
      description: Delete a receipt
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
      summary: Delete a receipt
      tags:
      - Receipts
    get:
      description: Find a receipt by id
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ReceiptResponse'
      summary: Find a receipt by id
      tags:
      - Receipts
    put:
      description: Update a receipt
      parameters:
      - description: Receipt id
        in: path
        name: id
        required: true
        type: string
      - description: Receipt
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/model.ReceiptRequest'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update a receipt
      tags:
      - Receipts
  /budget-control/api/v1/receipt/{year}/{month}:
    get:
      description: Find all receipts by Period
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: integer
      - description: Month
        in: path
        name: month
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ReceiptResponse'
            type: array
      summary: Find all receipts by Period
      tags:
      - Receipts
  /budget-control/api/v1/user:
    post:
      description: create a new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserRequest'
      tags:
      - User
swagger: "2.0"